tests/wpt/tests/css/css-box/margin-trim/computed-margin-values/grid-inline-start-item-negative-span.html

<!DOCTYPE html>
<html>
<head>
<link rel="author" href="mailto:sammy.gill@apple.com">
<link rel="help" href="https://drafts.csswg.org/css-box-4/#margin-trim-grid">
<meta name="assert" content="trimmed inline-start margins in grid should be reflected in computed style">
</head>
<style>
grid {
    display: grid;
    width: min-content;
    outline: 1px solid black;
    grid-template-columns: repeat(2, auto);
    margin-trim: inline-start;
}
item {
    display: block;
    width: 50px;
    height: 50px;
    margin-inline-start: 10px;
    background-color: green;
}
.negative-line-number {
    width: 50px;
    grid-row: 2;
    grid-column: -3;
    background-color: blue;
}
</style>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/check-layout-th.js"></script>
<body onload="checkLayout('grid > item')">
    <div id="target">
    <grid>
        <item data-expected-margin-left="0" class="negative-line-number"></item>
        <item data-expected-margin-left="0"></item>
    </grid>
    </div>
</body>
</html>
