third_party/blink/web_tests/external/wpt/css/css-box/margin-trim/computed-margin-values/grid-inline-start-item-negative-span-expected.txt

This is a testharness.js-based test.
[FAIL] grid > item 1
  assert_equals: \n<item data-expected-margin-left="0" class="negative-line-number"></item>\nmargin-left expected "0" but got "10"
[FAIL] grid > item 2
  assert_equals: \n<item data-expected-margin-left="0"></item>\nmargin-left expected "0" but got "10"
Harness: the test ran to completion.
